requirements.txt
absl-py==0.12.0
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
black==20.8b1
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
csvtomd==0.3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
dpkt==1.9.6
efficientnet==1.1.1
entrypoints==0.3
executing==0.6.0
fastapi==0.63.0
filelock==3.0.12
flake8==3.9.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gdown==3.12.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
huggingface-hub==0.0.8
idna==2.10
imageio==2.9.0
importlib-metadata==3.7.3
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
klib==0.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
nltk==3.6.2
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pathlib==1.0.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pip-chill==1.0.1
progressbar2==3.53.1
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.1
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.1
pydeck==0.6.1
pyflakes==2.3.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-multipart==0.0.5
python-utils==2.5.6
python-whois==0.7.3
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.3.17
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rich==10.4.0
rsa==4.7.2
sacremoses==0.0.45
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
starlette==0.13.6
streamlit==0.79.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.3.17
tldextract==3.1.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
transformers==4.6.1
typed-ast==1.4.2
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
uvicorn==0.13.4
validators==0.18.2
varname==0.6.5
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.2
zipp==3.4.1
